+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.8] - 2023-11-13
+
+### Added
+
+### Fixed
+
+### Changed
+
+- Change `supportedCountries` to `supportedCountriesIdDoc`.
+- Update README.md with updated function name.
+
+### Removed
+
 ## [0.5.7] - 2023-11-13
 
 ### Added